bug: fix compilation on stable
Move codegen-requiring code from `src/action.rs` into `src/action.in.rs`.
Add `src/action.in.rs` generation to `build.rs`.
Move codegen-requiring code from `src/camera.rs` into `src/camera.in.rs`.
Add `src/camera.in.rs` generation to `build.rs`.
Replace `create_config!` macro usage with a manually defined `Config` struct and
`Default` impl.
Add `src/localization.in.rs` generation to `build.rs`.
Replace `create_localization!` macro usage with a manually defined
`Localization` struct and `Default` impl.
Add `deserialize_from_file_or_default` function for deserializing
serde-compatible structs from either a file or their `Default` impl.
Use `deserialize_from_file_or_default` function to deserialize `Config` and
`Localization`.
Rename `read_file_to_json` function to `read_file_to_string`.
Fix double punctuation typo in `main.rs`.

Move code which does not require codegen from `framework/src/bindings.in.rs` to
`framework/src/bindings.rs`.
Remove `create_config!` and `create_localization!` macro definitions.

Move code which does not require codegen from `world/src/direction.in.rs` to
`world/src/direction.rs`.
