+++ framework/src/lib.rs
@@ -24,7 +24,6 @@ pub use util::RustcSerializeWrapper;
 pub mod backend;
 mod bindings;
 pub mod draw;
-mod macros;
 mod scene;
 mod scene_manager;
 mod util;
