+++ src/action.in.rs
@@ -1,0 +1,4 @@
+#[derive(Clone, Deserialize, Serialize)]
+pub enum Action {
+    Camera(CameraAction),
+}
