+++ src/action.rs
@@ -1,6 +1,7 @@
 use camera::CameraAction;
 
-#[derive(Clone, Deserialize, Serialize)]
-pub enum Action {
-    Camera(CameraAction),
-}
+#[cfg(feature = "nightly")]
+include!("action.in.rs");
+
+#[cfg(feature = "with-syntex")]
+include!(concat!(env!("OUT_DIR"), "/action.rs"));
