+++ src/camera.in.rs
@@ -1,0 +1,4 @@
+#[derive(Clone, Deserialize, Serialize)]
+pub enum CameraAction {
+    Move(Direction),
+}
