+++ src/localization.in.rs
@@ -1,14 +1,39 @@
-create_localization! {
-    colonize_window_title: String, "Colonize".to_owned(), "Colonize - Window title";
-    gamescene_welcome_text: String, "Welcome to Colonize!".to_owned(), "GameScene - Welcome text";
-    gamescene_debug_cursor: String, "Mouse Cursor".to_owned(), "GameScene - Debug - Cursor";
-    gamescene_debug_camera: String, "Camera".to_owned(), "GameScene - Debug - Camera";
-    gamescene_debug_chunk: String, "Chunk".to_owned(), "GameScene - Debug - Chunk";
-    internal_failed_to_build_window: String, "Failed to build window".to_owned(), "Internal - Failed to build window";
-    internal_failed_to_load_font: String, "Failed to load font".to_owned(), "Internal - Failed to load font message";
-    menuscene_singleplayer: String, "S)ingleplayer".to_owned(), "MenuScene - Menu option - Singleplayer";
-    menuscene_options: String, "O)ptions".to_owned(), "MenuScene - Menu option - Options";
-    menuscene_credits: String, "C)redits".to_owned(), "MenuScene - Menu option - Credits";
-    util_unit_millisecond: String, "ms".to_owned(), "Util - Unit - Millisecond";
-    util_unit_fps: String, "FPS".to_owned(), "Util - Unit - FPS";
+#[derive(Deserialize, Serialize)]
+pub struct Localization {
+    #[serde(default)]
+    /// Colonize - Window title
+    pub colonize_window_title: String,
+    #[serde(default)]
+    /// GameScene - Welcome text
+    pub gamescene_welcome_text: String,
+    #[serde(default)]
+    /// GameScene - Debug - Cursor
+    pub gamescene_debug_cursor: String,
+    #[serde(default)]
+    /// GameScene - Debug - Camera
+    pub gamescene_debug_camera: String,
+    #[serde(default)]
+    /// GameScene - Debug - Chunk
+    pub gamescene_debug_chunk: String,
+    #[serde(default)]
+    /// Internal - Failed to build window
+    pub internal_failed_to_build_window: String,
+    #[serde(default)]
+    /// Internal - Failed to load font message
+    pub internal_failed_to_load_font: String,
+    #[serde(default)]
+    /// MenuScene - Menu option - Singleplayer
+    pub menuscene_singleplayer: String,
+    #[serde(default)]
+    /// MenuScene - Menu option - Options
+    pub menuscene_options: String,
+    #[serde(default)]
+    /// MenuScene - Menu option - Credits
+    pub menuscene_credits: String,
+    #[serde(default)]
+    /// Util - Unit - Millisecond
+    pub util_unit_millisecond: String,
+    #[serde(default)]
+    /// Util - Unit - FPS
+    pub util_unit_fps: String,
 }
